refactoring nodejs-ssb
* Extracted pipelines
* Extracted caches
* Extracted job queues

still works.

related to #105
