+++ src/core/platforms/platforms.js
@@ -1,14 +1,17 @@
 const { NodeJsSSB } = require("./nodejs-ssb/ssb.js")
 const { GoSSB } = require("./go-ssb/ssb.js")
 
-
 const platforms = {
   "nodejs-ssb": NodeJsSSB,
   "go-ssb": GoSSB,
 }
 
 const setServerType = (serverType = "nodejs-ssb", keys, remote) => {
-  global.ssb = new platforms[serverType]()
+  if (platforms.hasOwnProperty(serverType)) {
+    global.ssb = new platforms[serverType]()
+  } else {
+    throw `unknown server type: ${serverType}`
+  }
 }
 
 global.setServerType = setServerType
